working dynamic image loading
