+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,18 +1,32 @@
-asset-manifest.json,1597935107429,27f9259a282fb29b5f6f89220a55707da20cfa977f04241fbe4dfb640ce27865
-index.html,1597935107429,666ccd0273a7fcc62e93bc241704df23b148c6390ad844ad5417c7830883a674
-manifest.json,1597895310183,5c997de1364b8be939319fa9209abd77f2caf7f8844999a9e2e9173f844e7840
-precache-manifest.43d77b66be9973850b94fe38f2199328.js,1597935107429,97d42c95c24ff3beec73d0efca417bea060b6de55872295bb585de7dd3e04b68
-logo512.png,1597895310182,ee7e2f3fdb8209c4b6fd7bef6ba50d1b9dba30a25bb5c3126df057e1cb6f5331
-service-worker.js,1597935107429,a766321d86d9ddd01fcdb5912b4ca052a9c201018fd44b2594baf6ecf40a4704
-logo192.png,1597895310182,3ee59515172ee198f3be375979df15ac5345183e656720a381b8872b2a39dc8b
-static/css/main.523f09c5.chunk.css,1597935107432,1532c68c7752d8131187c5fadbee2960d9eda20abb3a787ec4bd3c9ffecb001a
-static/js/2.dc7d7431.chunk.js.LICENSE.txt,1597935107434,898315b0062a4a359fb9ba91ae0c3cffdae64f31bfe083b551a42b3487609785
-static/css/main.523f09c5.chunk.css.map,1597935107435,65322ce08e5806aff183a46d901f441cddcf8db41a8ebc0f5b3b855692024d2a
-static/js/main.d13478e8.chunk.js,1597935107432,8b7e1e98591f6ac1797ca1b8918b4cbb69af5df84831d20612ae0350b4db71db
-static/js/runtime-main.6a16053d.js,1597935107434,0a1412ac0ad61f4c548aa0fafaf030af3027121ac3de5f0de640b27e983a7693
-static/js/runtime-main.6a16053d.js.map,1597935107435,51f25a8c20987929667231e4de70fa5cc0f6e98ee7f2b5afc110cfcba0f2e39b
-static/js/main.d13478e8.chunk.js.map,1597935107434,514d11023ab0e50398bb0acf5fc71ac377f5ab5d2cc5c33ccfc25ee8b5d1eac4
-robots.txt,1597895310183,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
 favicon.ico,1597895310181,d96ddbc4933b04e12c738ab39f469573143949ca2c39eda0a49d16f83d40c319
-static/js/2.dc7d7431.chunk.js,1597935107434,8911fafb6c41682487a8666e1d5e25de9796afebdea169042b36d902f4cbea7c
-static/js/2.dc7d7431.chunk.js.map,1597935107435,e76ded8301e9553a7fca6c747fe62f6fde564455830288021901b93845f478f8
+logo192.png,1597895310182,3ee59515172ee198f3be375979df15ac5345183e656720a381b8872b2a39dc8b
+logo512.png,1597895310182,ee7e2f3fdb8209c4b6fd7bef6ba50d1b9dba30a25bb5c3126df057e1cb6f5331
+manifest.json,1597895310183,5c997de1364b8be939319fa9209abd77f2caf7f8844999a9e2e9173f844e7840
+robots.txt,1597895310183,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
+imgs/portfolioSite.png,1597895310246,67854d5b8aba3e44e4c4de44f37f79402e42820947e9c2105e6dfef162f0ffdf
+imgs/2TapTimer/timerSelect.PNG,1597895310186,e0bcd8a99da5defe2940139c0924e184ed82b2fdc750de607d55498691b74554
+imgs/MLAirHockey/AirHockeyModel.jpeg,1597895310187,98826bc5be0e6e75192844392a80d504db52fdd172894030402f22621561d8dc
+imgs/MLAirHockey/hitrate.png,1597895310197,893ad5ee8f45b2214df17b7cb0d8c20b124c2a51e1a2676811b87f6dd2011a54
+imgs/MLAirHockey/MLSim.png,1597895310187,09efbccc2d03c2700c16a74176eef9924476c9f5d91513977cc3a6da73718b70
+imgs/MLAirHockey/Poster.png,1597895310196,3fb00c5484ebba67509c7155ee0033c6a8b5e3128135f72f52a194a2db04a19e
+imgs/MLAirHockey/stick.jpg,1597895310209,3f694bc8ac66eab11aa15169a2ffb65ec323aeebdeb15b5c500a8b7a853e8312
+imgs/MLAirHockey/stick_model.JPG,1597895310211,06a83ae006dcb2d4649b14cfd88d44c02accf831375eec286a0fec5e6763ed0c
+imgs/MLAirHockey/table.jpg,1597895310219,d0fbf996592aa23b0c26e7d369a82fc4398cddf6ce679d07eeeea59e01d6f4b1
+imgs/PillowFluffer/Overhead.png,1597895310224,3162021cd2a4be851a3e1e3d2d68b3e6987571f5c190894bd3fb2355c7a0eff9
+imgs/PillowFluffer/pov.png,1597895310228,f0716d09f4af3be9d1acade04d593d9307a4264a34b67ec096043169c51c399d
+imgs/PillowFluffer/unity.png,1597895310234,b8f341146e9d880be15257283e41ec276cdcbb585c8bcce00ada94b7a22f69c3
+asset-manifest.json,1597980557387,248f3306e1d82520dda8cd9ae468a31eaa43fc2aebb7a3ca8b3e663ddf9d9572
+precache-manifest.039c73b47d2f063b795b781c30b25952.js,1597980557386,c53ba611a643d33ee1ec4d29872bf730cc43d239a4d201c96f5b7a47521f17c2
+index.html,1597980557386,01baa9a16519f7c4a322745b0aa22f650456d6b29aab309d17325d39b13a0627
+service-worker.js,1597980557386,0eb1dace9b5a5e0acfd23578db8ac0e614b9a8ac3f265368fd5072dc015515db
+static/css/main.523f09c5.chunk.css,1597980557398,1532c68c7752d8131187c5fadbee2960d9eda20abb3a787ec4bd3c9ffecb001a
+static/js/2.c6b0c388.chunk.js.LICENSE.txt,1597980557401,898315b0062a4a359fb9ba91ae0c3cffdae64f31bfe083b551a42b3487609785
+static/css/main.523f09c5.chunk.css.map,1597980557401,65322ce08e5806aff183a46d901f441cddcf8db41a8ebc0f5b3b855692024d2a
+static/js/main.6e6fb389.chunk.js,1597980557401,e704ee5cfb99a110cb98edf6acb8f7189ae5e93abd77d50df174f859cf46ca02
+static/js/main.6e6fb389.chunk.js.map,1597980557401,216c5b5711cc63aefbd2100da4f8b0f7ff61c307ef0f25bfe3f9cab2c3613919
+static/js/runtime-main.6a16053d.js,1597980557401,0a1412ac0ad61f4c548aa0fafaf030af3027121ac3de5f0de640b27e983a7693
+static/js/runtime-main.6a16053d.js.map,1597980557403,51f25a8c20987929667231e4de70fa5cc0f6e98ee7f2b5afc110cfcba0f2e39b
+static/css/2.11829350.chunk.css,1597980557401,baba07a9b206346a7516773fa115d696c2c3f5ccfadf8ef1dd3fad8d158f5a30
+static/js/2.c6b0c388.chunk.js,1597980557401,40519b9fcf5b6159eb778215582110ea2b9277e54092578a0161db52fd5682f7
+static/css/2.11829350.chunk.css.map,1597980557401,ad662b160d71ee79e014dfad2134e186654b5c217d55b435a978cbdbef304bea
+static/js/2.c6b0c388.chunk.js.map,1597980557401,1e4010deda53171621c960823ab490c36def51de28ecfe8339757d63c0de4aff
