--- src/components/itemCard.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import './App.css';
-import { ReactComponent } from '*.svg';
-
-class ItemCard extends React.Component {
-    render(){
-        return (
-
-            <div>hello</div>
-        
-        );
-    }
-}
-
-export default ItemCard;
