+++ src/siteImages.json
@@ -1,0 +1,21 @@
+{
+    "2TapTimer":[
+        "/imgs/2TapTimer/timerSelect.PNG"
+    ],
+    "mlairhockey":[
+        "imgs/MLAirHockey/AirHockeyModel.jpeg",
+        "imgs/MLAirHockey/hitrate.png",
+        "imgs/MLAirHockey/MLSim.png",
+        "imgs/MLAirHockey/Poster.png",
+        "imgs/MLAirHockey/stick.jpg",
+        "imgs/MLAirHockey/stocl_model.JPG",
+        "imgs/MLAirHockey/table.jpg"
+    ],
+    "PillowFluffer":[
+        "imgs/PillowFluffer/Overhead.png",
+        "imgs/PillowFluffer/pov.png",
+        "imgs/PillowFluffer/unity.png"
+    ],    "portfolioSite": [
+        "imgs/portfolioSite/portfolioSite.png"
+    ]
+}
