--- src/tsconfig.js
@@ -1,3 +1,0 @@
-{
-
-}
